+++ Data/Repositories/Implements/LoanDetailRepository.cs
@@ -8,7 +8,7 @@ namespace Library.Data.Repositories.Implements;
 public class LoanDetailRepository(ApplicationDbContext appDbContext)
     : Repository<LoanDetail>(appDbContext), ILoanDetailRepository
 {
-    public async Task<LoanDetail> GetByLoanIdAndBookIdAsync(int loanId, int bookId)
+    public async Task<LoanDetail> GetByLoanIdAndBookIdAsync(long loanId, long bookId)
     {
         return await AppDbContext.LoanDetails.FirstAsync(loanDetail =>
             loanDetail.LoanId == loanId && loanDetail.BookId == bookId);
