+++ Data/Repositories/Interfaces/IBookRepository.cs
@@ -5,9 +5,9 @@ namespace Library.Data.Repositories.Interfaces;
 
 public interface IBookRepository : IRepository<Book>
 {
-    Task<List<Book>> GetBooksByAuthorAsync(int authorId);
+    Task<List<Book>> GetBooksByAuthorAsync(long authorId);
     Task<Book> AddAsync(BookDto bookDto);
     Task<List<Author>> GetAuthorsAsync(long bookId);
     Task<List<Category>> GetCategoriesAsync(long bookId);
-    Task<List<Book>> GetBooksByCategoryAsync(int categoryId);
+    Task<List<Book>> GetBooksByCategoryAsync(long categoryId);
 }
