+++ Data/Repositories/Interfaces/ILoanDetailRepository.cs
@@ -4,5 +4,5 @@ namespace Library.Data.Repositories.Interfaces;
 
 public interface ILoanDetailRepository : IRepository<LoanDetail>
 {
-    Task<LoanDetail> GetByLoanIdAndBookIdAsync(int loanId, int bookId);
+    Task<LoanDetail> GetByLoanIdAndBookIdAsync(long loanId, long bookId);
 }
