+++ Data/Repositories/Interfaces/IRepository.cs
@@ -9,8 +9,8 @@ namespace Library.Data.Repositories.Interfaces;
 public interface IRepository<T> where T : IEntity
 {
     Task<List<T>> GetAllAsync();
-    Task<T> GetByIdAsync(int id);
+    Task<T> GetByIdAsync(long id);
     Task<T> AddAsync(T entity);
-    Task<bool> UpdateAsync(int id, T entity);
-    Task<bool> DeleteAsync(int id);
+    Task<bool> UpdateAsync(long id, T entity);
+    Task<bool> DeleteAsync(long id);
 }
