+++ Services/Interfaces/IBookService.cs
@@ -9,8 +9,8 @@ namespace Library.Services.Interfaces;
  */
 public interface IBookService : IService<Book>
 {
-    List<Book> GetAllByAuthor(int authorId);
-    List<Book> GetAllByCategory(int categoryId);
+    List<Book> GetAllByAuthor(long authorId);
+    List<Book> GetAllByCategory(long categoryId);
     Book Add(BookDto bookDto);
     List<AuthorDto> GetAuthors(long bookId);
     List<CategoryDto> GetCategories(long bookId);
