+++ Services/Interfaces/ILoanDetailService.cs
@@ -4,5 +4,5 @@ namespace Library.Services.Interfaces;
 
 public interface ILoanDetailService : IService<LoanDetail>
 {
-    LoanDetail GetByLoanIdAndBookId(int loanId, int bookId);
+    LoanDetail GetByLoanIdAndBookId(long loanId, long bookId);
 }
